Fix tests broken by use of styled-icons library
The styled-icons library is un-transpiled, so when Jest encounters the
"export" statement, it throws a syntax error. I tried to add
styled-icons to Jest's whitelist for transpiling node_module packages,
but it didn't work. So I made a mock for it instead.

Incidentally I had to rewrite some of the tests for MenuListItem because
I changed the the name of the target subcomponents.