Fix tests for AddItem broken by StyledComponents
The tests find the relevant DOM elements to simulate clicks and changes.
However, when the elements were styled using StyledComponents, the id's
were applied to all the child components generated by StyledComponents.
This caused Jest to throw an error when it found 3 nodes instead of 1.
A workaround was implemented by adding the DOM element to the selector.